requirements.txt
wheel
neovim
pylama
black
ipython
livemark
PyGithub
python-dotenv
